Message isolated context directly in runserver
