+++ mountaineer/development/isolation.py
@@ -18,86 +18,24 @@ from mountaineer.controllers.exception_controller import (
 )
 from mountaineer.development.hotreload import HotReloader
 from mountaineer.development.messages import (
-    AsyncMessageBroker,
     BootupMessage,
     BuildJsMessage,
     BuildUseServerMessage,
-    CaptureLogsSuccessResponse,
     ErrorResponse,
     IsolatedMessageBase,
-    ReloadModulesMessage,
-    ReloadResponseError,
-    ReloadResponseSuccess,
-    RestartServerMessage,
-    ShutdownMessage,
-    StartCaptureLogsMessage,
-    StopCaptureLogsMessage,
     SuccessResponse,
 )
+from mountaineer.development.messages_broker import AsyncMessageBroker
 from mountaineer.development.uvicorn import UvicornThread
 from mountaineer.logging import setup_internal_logger
 
 LOGGER = setup_internal_logger(__name__)
 
 
-class IsolatedAppContext(Process):
+class IsolatedAppContext:
     """
-    Isolated process that manages the app controller, server, and compilation.
+    Manages the app controller, server, and compilation.
 
-    This class runs the application in a separate process, providing a clean
-    environment for module reloading and compilation without affecting the main process.
-
-    It communicates with the main process through a message broker, handling various
-    operations like module reloading, server restarts, and JS compilation. This isolation
-    enables Mountaineer's hot-reloading capabilities by allowing code to be reloaded
-    without restarting the entire application.
-
-    ```python {{sticky: True}}
-    import asyncio
-    from pathlib import Path
-    from mountaineer.development.messages import AsyncMessageBroker, BootupMessage, ShutdownMessage
-    from mountaineer.development.isolation import IsolatedAppContext
-
-    # Create a message broker for communication
-    broker = AsyncMessageBroker()
-    broker.start()
-
-    # Initialize the isolated context
-    isolated_app = IsolatedAppContext(
-        package="my_app",
-        package_path=Path("./my_app"),
-        module_name="my_app.main",
-        controller_name="app",
-        host="127.0.0.1",
-        port=8000,
-        live_reload_port=3001,
-        message_broker=broker
-    )
-
-    # Start the isolated process
-    isolated_app.start()
-
-    try:
-        # Send bootup message to initialize the app
-        await broker.send_message(BootupMessage())
-
-        print("App started successfully in isolated process!")
-
-        # Wait for some time (in a real app, this would be until shutdown is needed)
-        await asyncio.sleep(5)
-
-    finally:
-        # Send shutdown message and wait for it to complete
-        await broker.send_message(ShutdownMessage())
-
-        # Wait for the process to terminate
-        isolated_app.join(timeout=2)
-        if isolated_app.is_alive():
-            isolated_app.terminate()
-
-        # Clean up the broker
-        await broker.stop()
-    ```
     """
 
     def __init__(
@@ -109,7 +47,6 @@ class IsolatedAppContext(Process):
         host: str | None,
         port: int,
         live_reload_port: int | None,
-        message_broker: AsyncMessageBroker[IsolatedMessageBase[Any]],
     ):
         """
         Initialize an isolated application context in a separate process.
@@ -131,38 +68,36 @@ class IsolatedAppContext(Process):
         self.host = host
         self.port = port
         self.live_reload_port = live_reload_port
-        self.message_broker = message_broker
         self.webservice_thread: UvicornThread | None = None
 
         self.app_controller: AppController | None = None
 
         self.js_compiler: APIBuilder | None = None
         self.app_compiler: ClientCompiler | None = None
-        self.hot_reloader: HotReloader | None = None
 
-        # Log capture state
-        self._stdout_capture: StringIO | None = None
-        self._stderr_capture: StringIO | None = None
-        self._stdout_redirect: redirect_stdout | None = None
-        self._stderr_redirect: redirect_stderr | None = None
+    @classmethod
+    def from_webcontroller(
+        cls,
+        webcontroller: str,
+        host: str | None = None,
+        port: int | None = None,
+        live_reload_port: int | None = None,
+    ):
+        package = webcontroller.split(".")[0]
+        module_name = webcontroller.split(":")[0]
+        controller_name = webcontroller.split(":")[1]
 
-    def run(self):
-        """
-        Main process entry point that runs when the process starts.
+        return cls(
+            package=package,
+            package_path=Path(package.replace(".", "/")),
+            module_name=module_name,
+            controller_name=controller_name,
+            host=host,
+            port=port,
+            live_reload_port=live_reload_port,
+        )
 
-        This method is called automatically when the process is started with start().
-        It runs the async event loop and handles exceptions gracefully.
-        """
-        try:
-            asyncio.run(self.run_async())
-        except (KeyboardInterrupt, asyncio.CancelledError):
-            pass
-        except Exception as e:
-            LOGGER.error(f"Isolated app context failed: {e}", exc_info=True)
-
-        LOGGER.debug("Isolated app shutdown complete")
-
-    async def run_async(self):
+    async def run_async(self, broker: AsyncMessageBroker):
         """
         Asynchronous main loop that processes messages from the broker.
 
@@ -178,41 +113,27 @@ class IsolatedAppContext(Process):
 
             # Process messages until shutdown
             while True:
-                message_id, message = self.message_broker.message_queue.get()
+                message_id, message = broker.message_queue.get()
                 LOGGER.debug(f"[IsolatedAppContext] Got message: {message}")
 
                 try:
                     response: SuccessResponse | ErrorResponse
                     if isinstance(message, BootupMessage):
                         response = await self.handle_bootstrap()
-                    elif isinstance(message, ReloadModulesMessage):
-                        response = await self.handle_module_reload(message.module_names)
-                    elif isinstance(message, RestartServerMessage):
-                        response = await self.handle_restart_server()
                     elif isinstance(message, BuildJsMessage):
                         response = await self.handle_js_build(message.updated_js)
                     elif isinstance(message, BuildUseServerMessage):
                         response = await self.handle_build_use_server()
-                    elif isinstance(message, StartCaptureLogsMessage):
-                        response = await self.handle_start_capture_logs()
-                    elif isinstance(message, StopCaptureLogsMessage):
-                        response = await self.handle_stop_capture_logs()
-                    elif isinstance(message, ShutdownMessage):
-                        # Send immediately before breaking out of the loop, which should
-                        # trigger a shutdown of this process
-                        response = await self.handle_shutdown()
-                        self.message_broker.response_queue.put((message_id, response))
-                        break
                     else:
                         LOGGER.error(f"Invalid message type: {type(message)} {message}")
                         continue
-                    self.message_broker.response_queue.put((message_id, response))
+                    broker.response_queue.put((message_id, response))
                 except Exception as e:
                     LOGGER.info(
                         f"Isolated app context failed: {e}, continuing...",
                         exc_info=True,
                     )
-                    self.message_broker.response_queue.put(
+                    broker.response_queue.put(
                         (
                             message_id,
                             ErrorResponse(
@@ -248,20 +169,6 @@ class IsolatedAppContext(Process):
 
         return response
 
-    async def handle_restart_server(self):
-        """
-        Restart the web server with the current app controller.
-
-        This reloads the web service and starts a new server instance.
-        Used when the application structure has changed but doesn't need full reinitialization.
-
-        :return: Success response on successful restart
-        """
-        # Restart the server with new controller
-        self.load_webservice()
-        await self.start_server()
-        return SuccessResponse()
-
     async def handle_build_use_server(self):
         """
         Build the useServer support files for client-side React Server Components.
@@ -274,42 +181,6 @@ class IsolatedAppContext(Process):
 
         await self.js_compiler.build_use_server()
         return SuccessResponse()
-
-    async def handle_module_reload(self, module_names: list[str]):
-        """
-        Reload specified Python modules in the isolated context.
-
-        Uses the HotReloader to safely reload modules and their dependencies.
-
-        :param module_names: List of module names to reload
-        :return: Success or error response with details about reloaded modules and restart needs
-        """
-        needs_restart = True
-        reloaded: list[str] = []
-
-        try:
-            if self.hot_reloader is None:
-                raise ValueError("Hot reloader not initialized")
-
-            # Get the list of modules to reload from the hot reloader
-            reload_status = self.hot_reloader.reload_modules(module_names)
-
-            if reload_status.error:
-                reloaded = reload_status.reloaded_modules
-                raise reload_status.error
-
-            return ReloadResponseSuccess(
-                reloaded=reload_status.reloaded_modules,
-                needs_restart=reload_status.needs_restart,
-            )
-        except Exception as e:
-            LOGGER.debug(f"Failed to reload modules: {e}", exc_info=True)
-            return ReloadResponseError(
-                reloaded=reloaded,
-                needs_restart=needs_restart,
-                exception=str(e),
-                traceback="".join(format_exception(e)),
-            )
 
     async def handle_js_build(self, updated_js: list[Path] | None = None):
         """
@@ -330,65 +201,6 @@ class IsolatedAppContext(Process):
         for path in updated_js or []:
             self.app_controller.invalidate_view(path)
 
-        return SuccessResponse()
-
-    async def handle_start_capture_logs(self) -> SuccessResponse:
-        """
-        Start capturing stdout and stderr during module reload.
-
-        This method activates the context managers and saves the captures to the instance.
-
-        :return: Success response when capture is started
-        """
-        self._stdout_capture = StringIO()
-        self._stderr_capture = StringIO()
-        self._stdout_redirect = redirect_stdout(self._stdout_capture)
-        self._stderr_redirect = redirect_stderr(self._stderr_capture)
-        self._stdout_redirect.__enter__()
-        self._stderr_redirect.__enter__()
-
-        return SuccessResponse()
-
-    async def handle_stop_capture_logs(self) -> CaptureLogsSuccessResponse:
-        """
-        End capturing stdout and stderr, restore standard streams, and return the captured content.
-
-        :return: Success response containing the captured stdout and stderr
-        :raises RuntimeError: If capturing was not started before stopping
-        """
-        if (
-            not self._stdout_capture
-            or not self._stderr_capture
-            or not self._stdout_redirect
-            or not self._stderr_redirect
-        ):
-            raise RuntimeError("Cannot end capture logs before starting")
-
-        self._stdout_redirect.__exit__(None, None, None)
-        self._stderr_redirect.__exit__(None, None, None)
-
-        stdout_capture = self._stdout_capture
-        stderr_capture = self._stderr_capture
-
-        # Clear the state
-        self._stdout_capture = None
-        self._stderr_capture = None
-        self._stdout_redirect = None
-        self._stderr_redirect = None
-
-        return CaptureLogsSuccessResponse(
-            captured_logs=stdout_capture.getvalue(),
-            captured_errors=stderr_capture.getvalue(),
-        )
-
-    async def handle_shutdown(self):
-        """
-        Handle shutdown of the isolated context by stopping the web server.
-
-        :return: Success response when shutdown is complete
-        """
-        if self.webservice_thread:
-            await self.webservice_thread.astop()
         return SuccessResponse()
 
     #
@@ -413,13 +225,6 @@ class IsolatedAppContext(Process):
 
         if self.app_controller is None:
             raise ValueError("App controller not initialized")
-
-        # Initialize hot reloader
-        self.hot_reloader = HotReloader(
-            root_package=self.package,
-            package_path=self.package_path,
-            entrypoint=self.module_name,
-        )
 
         # Mount exceptions
         self.exception_controller = ExceptionController()
