+++ pyproject.toml
@@ -20,9 +20,9 @@ websockets = "<14.0"
 
 # TODO: Move to optional dependencies, since release configurations
 # will not need these.
-watchfiles = "^0.21.0"
+watchfiles = "^1.0.0"
 rich = "^13.7.1"
-firehot = "^0.1.0"
+firehot = "^0.2.0"
 
 [tool.poetry.group.dev.dependencies]
 mypy = "^1.8.0"
@@ -95,11 +95,11 @@ dependencies = [
     "click>=8.1.7,<9.0.0",
     "uvicorn[standard]>=0.27.0.post1,<1.0.0",
     "packaging>=23.2",
-    "watchfiles>=0.21.0,<1.0.0",
+    "watchfiles>=1.0.0,<2.0.0",
     "pydantic-settings>=2.1.0,<3.0.0",
     "rich>=13.7.1,<14.0.0",
     "websockets<14.0",
-    "firehot>=0.1.0,<1.0.0",
+    "firehot>=0.2.0,<1.0.0",
 ]
 
 # Excluded in sdist packaging
